+++ castero/datafile.py
@@ -23,7 +23,7 @@ class DataFile:
                               os.path.join(HOME, '.local', 'share'))
     CONFIG_DIR = os.path.join(XDG_CONFIG_HOME, castero.__title__)
     DATA_DIR = os.path.join(XDG_DATA_HOME, castero.__title__)
-    DOWNLOADED_DIR = os.path.join(DATA_DIR, "downloaded")
+    DEFAULT_DOWNLOADED_DIR = os.path.join(DATA_DIR, "downloaded")
 
     def __init__(self, path, default_path) -> None:
         """Initializes the object.